Issue #136 refactor: postservice 패키지 분리
